Move exampleUUID to SupplierTestHelper
