+++ java/clients/src/test/java/sleeper/clients/testutil/ClientTestUtils.java
@@ -16,13 +16,10 @@
 package sleeper.clients.testutil;
 
 import com.google.common.io.CharStreams;
-import org.apache.commons.lang3.StringUtils;
 
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.Reader;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class ClientTestUtils {
 
@@ -33,27 +30,5 @@ public class ClientTestUtils {
         try (Reader reader = new InputStreamReader(ClientTestUtils.class.getClassLoader().getResourceAsStream(path))) {
             return CharStreams.toString(reader);
         }
-    }
-
-    public static String exampleUUID(String start, Object uuidChar) {
-        if (start.length() > 8) {
-            throw new IllegalArgumentException("Start must be shorter than 8 characters: " + start);
-        }
-        char character = characterToRepeat(uuidChar);
-        return start + IntStream.of(8 - start.length(), 4, 4, 4, 12)
-                .mapToObj(size -> uuidPart(size, character))
-                .collect(Collectors.joining("-"));
-    }
-
-    private static String uuidPart(int size, char character) {
-        return StringUtils.repeat(character, size);
-    }
-
-    private static char characterToRepeat(Object charObj) {
-        String charStr = charObj.toString();
-        if (charStr.length() != 1) {
-            throw new IllegalArgumentException("Character to repeat must be a single character: " + charStr);
-        }
-        return charStr.charAt(0);
     }
 }
