+++ java/core/src/test/java/sleeper/core/testutils/SupplierTestHelper.java
@@ -15,10 +15,14 @@
  */
 package sleeper.core.testutils;
 
+import org.apache.commons.lang3.StringUtils;
+
 import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 /**
@@ -38,6 +42,19 @@ public class SupplierTestHelper {
      */
     public static Supplier<String> fixIds(String... ids) {
         return List.of(ids).iterator()::next;
+    }
+
+    /**
+     * Creates a supplier of IDs that would usually be generated with UUID.randomUUID. The IDs will be generated with
+     * numbers from 1 to 9, and then throw an exception when these are exhausted.
+     *
+     * @param  prefix the string to start each ID, at most 8 characters
+     * @return        the supplier
+     */
+    public static Supplier<String> supplyNumberedIdsWithPrefix(String prefix) {
+        return IntStream.range(1, 10)
+                .mapToObj(i -> exampleUUID(prefix, i))
+                .iterator()::next;
     }
 
     /**
@@ -73,4 +90,33 @@ public class SupplierTestHelper {
         return List.of(times).iterator()::next;
     }
 
+    /**
+     * Creates an example UUID with the first few characters replaced with a string, and the rest set to a given
+     * character.
+     *
+     * @param  start    the string for the start of the UUID, at most 8 characters
+     * @param  uuidChar the character to repeat for the rest of the UUID
+     * @return          the example UUID as a string
+     */
+    public static String exampleUUID(String start, Object uuidChar) {
+        if (start.length() > 8) {
+            throw new IllegalArgumentException("Start must be shorter than 8 characters: " + start);
+        }
+        char character = characterToRepeat(uuidChar);
+        return start + IntStream.of(8 - start.length(), 4, 4, 4, 12)
+                .mapToObj(size -> uuidPart(size, character))
+                .collect(Collectors.joining("-"));
+    }
+
+    private static String uuidPart(int size, char character) {
+        return StringUtils.repeat(character, size);
+    }
+
+    private static char characterToRepeat(Object charObj) {
+        String charStr = charObj.toString();
+        if (charStr.length() != 1) {
+            throw new IllegalArgumentException("Character to repeat must be a single character: " + charStr);
+        }
+        return charStr.charAt(0);
+    }
 }
